Fix overused namespace ModelSerializer
Our model pulp.app.models.base.ModelSerializer has this ancestry:
    `rest_framework.serializers.HyperlinkedModelSerializer`,
    `rest_framework.serializers.ModelSerializer`

This does not cause problems with the code, but may present some
confusion to developers who are used to the Django Rest Framework
because the idiomatic way to create a Serializer is to use the
`rest_framework.serializers.ModelSerializer` as the base class. If
someone used this in the code rather than our serializer, it would work
most of the time and might cause subtle problems. Additionally, it would
be used `serializers.ModelSerializer`. This could be identical to how you
would invoke `pulp.app.serializers.ModelSerializer`.